Create myshell.c
Write a program called my_shell, which is a step towards a simple command shell to process commands as follows:  print a prompt (e.g., “% “) then read a command on one line, then evaluate it.  A command can have the following form:
commandName [ argumentList ] [ IORedirect ] [ IORedirect ]
Where an argumentList is a list of zero or more words (no patterns allowed), IORedirect is either “< inFile” or “> outFile” or both where inFile and outFile are valid Unix file names. You will use versions of the system calls fork(), open(), dup2(), exec(), and wait() for this assignment.
You should test my_shell with these commands: (note # to end of line is a comment to you to help you understand what the command is supposed to do. existingInputFile means you have a file in your current directory that contains some text. newOutputFile means it will be created by the execution of that command.)
% /usr/bin/tee newOutputFile  # input from and output to the terminal
% /usr/bin/tee newOutputFile < existingInputFile
% /usr/bin/tee newOutputFile1 < existingInputFile > newOutputFile2

Command shell. Add pipes and background jobs. Check out this diagram of how pipe might be handled.  I also wrote an example program that implements the command cat < in | cat > out in this file. Only the first command in a pipeline can have its input redirected to come from a file and only the last command in a pipeline can have its output redirected to a file, so you can execute these commands:
% tee newOutputFile1 > newOutputFile2 < existingInputFile &
% cat < existingInputFile | tee newOutputFile1 > newOutputFile2
% cat < existingInputFile | tee newOutputFile1 > newOutputFile2 &
% cat < existingInputFile | tr A-Z a-z | tee newOutputFile1 | tr a-z A-Z > newOutputFile2
% cat < existingInputFile | tr A-Z a-z | tee newOutputFile1 | tr a-z A-Z > newOutputFile2 &