Create pool_server.c
Thread Pool Server 2 
Write a multi-threaded server that delivers files to clients above, but use a synchronized queue between the thread that reads requests from the clients and a pool of 5 threads that process client requests. Note all the syncyronization should happen in the queue.  Here is one implementation of the queue using pthreads and two mutex.  The important functions are putOnQ and getOffQ.